+++ azsettings/env.go
@@ -27,6 +27,8 @@ const (
 	UserIdentityAssertion                  = "GFAZPL_USER_IDENTITY_ASSERTION"
 	UserIdentityFallbackCredentialsEnabled = "GFAZPL_USER_IDENTITY_FALLBACK_SERVICE_CREDENTIALS_ENABLED"
 
+	AzureEntraPasswordCredentialsEnabled = "GFAZPL_AZURE_ENTRA_PASSWORD_CREDENTIALS_ENABLED"
+
 	// Pre Grafana 9.x variables
 	fallbackAzureCloud              = "AZURE_CLOUD"
 	fallbackManagedIdentityEnabled  = "AZURE_MANAGED_IDENTITY_ENABLED"
@@ -111,6 +113,14 @@ func ReadFromEnv() (*AzureSettings, error) {
 			UsernameAssertion: usernameAssertion,
 		}
 		azureSettings.UserIdentityFallbackCredentialsEnabled = serviceCredentialsFallback
+	}
+
+	// Client Password Credentials auth
+	if AzureEntraPasswordCredentialsEnabled, err := envutil.GetBoolOrDefault(AzureEntraPasswordCredentialsEnabled, false); err != nil {
+		err = fmt.Errorf("invalid Azure configuration: %w", err)
+		return nil, err
+	} else {
+		azureSettings.AzureEntraPasswordCredentialsEnabled = AzureEntraPasswordCredentialsEnabled
 	}
 
 	return azureSettings, nil